+++ shadow.go
@@ -151,13 +151,13 @@ func (shadow *viewShadowData) cssStyle(buffer *strings.Builder, session Session,
 		buffer.WriteString("inset ")
 	}
 
-	buffer.WriteString(offsetX.cssString("0"))
+	buffer.WriteString(offsetX.cssString("0", session))
 	buffer.WriteByte(' ')
-	buffer.WriteString(offsetY.cssString("0"))
+	buffer.WriteString(offsetY.cssString("0", session))
 	buffer.WriteByte(' ')
-	buffer.WriteString(blurRadius.cssString("0"))
+	buffer.WriteString(blurRadius.cssString("0", session))
 	buffer.WriteByte(' ')
-	buffer.WriteString(spreadRadius.cssString("0"))
+	buffer.WriteString(spreadRadius.cssString("0", session))
 	buffer.WriteByte(' ')
 	buffer.WriteString(color.cssString())
 	return true
@@ -177,11 +177,11 @@ func (shadow *viewShadowData) cssTextStyle(buffer *strings.Builder, session Sess
 	}
 
 	buffer.WriteString(lead)
-	buffer.WriteString(offsetX.cssString("0"))
+	buffer.WriteString(offsetX.cssString("0", session))
 	buffer.WriteByte(' ')
-	buffer.WriteString(offsetY.cssString("0"))
+	buffer.WriteString(offsetY.cssString("0", session))
 	buffer.WriteByte(' ')
-	buffer.WriteString(blurRadius.cssString("0"))
+	buffer.WriteString(blurRadius.cssString("0", session))
 	buffer.WriteByte(' ')
 	buffer.WriteString(color.cssString())
 	return true